keeping example up to date

+++ src/main.rs
@@ -90,7 +90,7 @@ fn main() {
 
     //
 
-    sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLContextProfileMask, sdl2::video::GLProfile::GLCoreProfile as isize);
+    sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLContextProfileMask, sdl2::video::GLProfile::GLCoreProfile as i32);
     sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLContextMajorVersion, 3);
     sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLContextMinorVersion, 3);
     sdl2::video::gl_set_attribute(sdl2::video::GLAttr::GLDoubleBuffer, 1);
